requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
faiss-cpu==1.9.0.post1
filelock==3.16.1
fsspec==2024.9.0
ftfy==6.3.1
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.3
open_clip_torch==2.29.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==3.20.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.1
timm==1.0.11
tokenizers==0.13.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.30.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
